refactor: use index for re-export
